Deleted unnecessary things

+++ src/main/kotlin/com/itmo/microservices/demo/orders/api/model/OrderDto.kt
@@ -8,7 +8,7 @@ data class OrderDto (
     val id: UUID,
     val timeCreated: Long,
     val status: OrderStatus,
-    val itemsMap: Map<UUID, Long>,
+    val itemsMap: Map<UUID, Int>,
     val deliveryDuration: Int?,
     val paymentHistory: List<PaymentLogRecord>
 )
